feat: add header group show
