+++ src/node/plugin.ts
@@ -9,7 +9,6 @@ import babelPluginIsland from './babel-plugin-island';
 import { ISLAND_JSX_RUNTIME_PATH } from './constants/index';
 import pluginUnocss from 'unocss/vite';
 import unocssOptions from './unocss.config';
-// import pluginInspect from 'vite-plugin-inspect';
 
 export async function createIslandPlugins(
   config: SiteConfig,
@@ -18,10 +17,6 @@ export async function createIslandPlugins(
 ): Promise<PluginOption[]> {
   return [
     pluginUnocss(unocssOptions),
-    // pluginInspect({
-    //   dev: false,
-    //   build: true
-    // }),
     // Md(x) compile
     await pluginMdx(config, isServer),
     // For island internal use
