+++ src/theme-default/components/Search/index.tsx
@@ -4,72 +4,13 @@ import { ComponentPropsWithIsland } from '../../../shared/types/index';
 import SearchSvg from './icons/search.svg';
 import LoadingSvg from './icons/loading.svg';
 import { throttle } from 'lodash-es';
+import { SuggestionContent } from './Suggestion';
 
 const KEY_CODE = {
   ARROW_UP: 'ArrowUp',
   ARROW_DOWN: 'ArrowDown',
   ENTER: 'Enter'
 };
-
-function SuggestionContent(props: {
-  suggestion: MatchResultItem;
-  query: string;
-  isCurrent: boolean;
-}) {
-  const { suggestion, query } = props;
-  const renderHeaderMatch = () => {
-    if (suggestion.type === 'header') {
-      const { header, headerHighlightIndex } = suggestion;
-      const headerPrefix = header.slice(0, headerHighlightIndex);
-      const headerSuffix = header.slice(headerHighlightIndex + query.length);
-      return (
-        <div font="medium">
-          <span>{headerPrefix}</span>
-          <span bg="brand-light" p="y-0.4 x-0.8" rounded="md" text="text-1">
-            {query}
-          </span>
-          <span>{headerSuffix}</span>
-        </div>
-      );
-    } else {
-      return <div font="medium">{suggestion.header}</div>;
-    }
-  };
-  const renderStatementMatch = () => {
-    if (suggestion.type !== 'content') {
-      return;
-    }
-    const { statementHighlightIndex, statement } = suggestion;
-    const statementPrefix = statement.slice(0, statementHighlightIndex);
-    const statementSuffix = statement.slice(
-      statementHighlightIndex + query.length
-    );
-    return (
-      <div font="normal" text="sm gray-light" w="100%">
-        <span>{statementPrefix}</span>
-        <span bg="brand-light" p="y-0.4 x-0.8" rounded="md" text="[#000]">
-          {query}
-        </span>
-        <span>{statementSuffix}</span>
-      </div>
-    );
-  };
-  return (
-    <div
-      border-1=""
-      table-cell=""
-      p="x-3 y-2"
-      hover="bg-[#f3f4f5]"
-      className={`border-right-none ${props.isCurrent ? 'bg-[#f3f4f5]' : ''}`}
-      transition="bg duration-200"
-    >
-      <div font="medium" text="sm">
-        {renderHeaderMatch()}
-      </div>
-      {suggestion.type === 'content' && renderStatementMatch()}
-    </div>
-  );
-}
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 export function Search(
