+++ src/theme-default/logic/index.ts
@@ -1,31 +1,3 @@
-export const isProduction = () => import.meta.env.PROD;
-
-export function addLeadingSlash(url: string) {
-  return url.charAt(0) === '/' ? url : '/' + url;
-}
-
-export function removeTrailingSlash(url: string) {
-  return url.charAt(url.length - 1) === '/' ? url.slice(0, -1) : url;
-}
-
-export function normalizeHref(url?: string) {
-  if (!url) {
-    return '/';
-  }
-  if (!isProduction() || url.startsWith('http')) {
-    return url;
-  }
-
-  let suffix = '';
-  if (!import.meta.env.ENABLE_SPA) {
-    suffix += '.html';
-    if (url.endsWith('/')) {
-      suffix = 'index' + suffix;
-    }
-  }
-  return addLeadingSlash(`${url}${suffix}`);
-}
-
 export { usePrevNextPage } from './usePrevNextPage';
 export { useEditLink } from './useEditLink';
 export { useSidebarData } from './useSidebarData';
@@ -33,3 +5,4 @@ export { useLocaleSiteData } from './useLocaleSiteData';
 export { setupEffects, bindingAsideScroll } from './sideEffects';
 export { setupCopyCodeButton } from './copyCode';
 export { PageSearcher } from './search';
+export * from './utils';
